- renames root_nav_graph - moves MainFragment - renames color
