+++ app/src/main/res/layout/activity_root.xml
@@ -16,6 +16,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:navGraph="@navigation/root_nav_graph" />
+        app:navGraph="@navigation/auth_nav_graph" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
