+++ app/src/main/res/layout/fragment_main.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".feature.MainFragment">
+    tools:context=".base.MainFragment">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/mainFragmentContainerView"
