+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -2,7 +2,7 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/internal_nav_graph.xml"
+    android:id="@+id/internal_nav_graph"
     app:startDestination="@id/feedFragment">
 
     <fragment
