+++ uikit/src/main/res/values/colors.xml
@@ -19,10 +19,9 @@
     <color name="butt_text_primary">#FFFFFF</color>
     <color name="butt_text_secondary">#1B58F1</color>
   
-    <color name="stroke">#5A45FE</color>
+    <color name="accent_secondary">#5A45FE</color>
 
     <color name="like_color">#FF3030</color>
     <color name="like_background_color">#1FFF3030</color>
 
-    <color name="accent_secondary">#5A45FE</color>
 </resources>
